refactor: move default safe transaction to utils
