+++ src/utils/safes/index.ts
@@ -138,3 +138,19 @@ export const nftNameContainsTokenId = (tokenName: string): boolean => {
 
   return false;
 };
+
+export const defaultTransaction = {
+  // type SafeTransaction
+  transactionType: '',
+  tokenData: null,
+  amount: null,
+  rawAmount: null,
+  recipient: undefined,
+  data: '',
+  contract: null,
+  abi: '',
+  contractFunction: '',
+  nft: null,
+  nftData: null,
+  functionParamTypes: null,
+};
